[#23] ChatDetail 기초 틀 설계
