+++ app/src/main/java/com/wd/woodong2/data/model/MessageItemResponse.kt
@@ -1,0 +1,14 @@
+package com.wd.woodong2.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class MessageItemsResponse(
+    val messageItems: List<MessageResponse>?
+)
+
+data class MessageResponse(
+    @SerializedName("id") val id: String?,
+    @SerializedName("message") val message: String?,
+    @SerializedName("timestamp") val timestamp: String?,
+    @SerializedName("senderId") val senderId: String?,
+)
